Fix NullPointerException in src/App.java (line 4) - Validated with compilation and tests

+++ src/App.java
@@ -1,7 +1,7 @@
 public class App {
     public static void main(String[] args) {
         String message = null;
-        Integer l = (message != null) ? message.length() : 0;
+Integer l = (message != null) ? (Integer) message.length() : 0;
 
         System.out.println(l);
     }
